Fixed warning in CubemapProbe Made PccGridPlacement live in the stack, to make it more clear it's not necessary to keep it alive
--HG--
branch : v2-2-irradiance-field

+++ Components/Hlms/Pbs/src/Cubemaps/OgreCubemapProbe.cpp
@@ -59,7 +59,7 @@ namespace Ogre
         mInvOrientation( Matrix3::IDENTITY ),
         mProbeShape( Aabb::BOX_ZERO ),
         mTexture( 0 ),
-        mCubemapArrayIdx( std::numeric_limits<uint32>::max() ),
+        mCubemapArrayIdx( std::numeric_limits<uint16>::max() ),
         mMsaa( 1u ),
         mWorkspaceMipmapsExecMask( 0x01 ),
         mClearWorkspace( 0 ),
